update sdk and bitcoinjs-lib
Summary:
update sdk and bitcoinjs-lib

Update the SDK to the new version 0.10.0 and update bitcoinjs-lib to the new
version. Other than fixing minor usages of bitcoinjs-lib, we make one other
change which is to use the name "bitcoin" for the main bitcoin network, rather
than "prod", which is what the new version of the SDK expects. We also add
istanbul test coverage. Note that the methods setupProxy and validateTxOutputs
were either modified or use bitcoinjs-lib, but are not covered by the tests,
and so need to be tested by hand.

Reviewers: ben, benchan

Reviewed By: benchan

Differential Revision: https://phabricator.bitgo.com/D865